fix bug on allowing user to evaluate an idea when it has no comments + allow user to delete one of his idea + select the right application field of an idea by default in the edit view

+++ projetAPI01/src/com/utc/projetAPI01/dao/IdeaDAOImpl.java
@@ -156,11 +156,13 @@ public class IdeaDAOImpl extends DAOAbstract<Idea>{
 		}
 		
 		CommentsDAOImpl commentsDao = new CommentsDAOImpl();
-		Iterator<Comments> itCo = idea.getComments().iterator();
-		while(itCo.hasNext())
-		{
-			Comments comment = itCo.next();
-			commentsDao.delete(comment);
+		if(idea.getComments() != null){
+			Iterator<Comments> itCo = idea.getComments().iterator();
+			while(itCo.hasNext())
+			{
+				Comments comment = itCo.next();
+				commentsDao.delete(comment);
+			}
 		}
 		
 		super.delete(idea);
